packages/backend/requirements.txt
# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0

# Database
sqlmodel>=0.0.14

# Authentication
PyJWT>=2.8.0

# Supabase
supabase>=2.3.4
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For async testing with FastAPI
pytest-asyncio>=0.23.5  # For async test support

# Development
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0